+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+ - Implement `From<&String>` for `DimensionName`
+
 ## [0.15.0] - 2024-07-07
 
 ### Highlights